hw04/pyproject.toml
# pyproject.toml.jinja

[project]
name = "hw04"
version = "0.1.0"
description = "CIFAR10. CIFAR100."
readme = "README.md"
authors = [
    { name = "Lennart Vetter.", email = "my@email" }
]
requires-python = ">=3.13"
dependencies = [
    "structlog",
    "rich",
    "numpy",
    "jax[cpu]>=0.7.2", 
    "jaxlib>=0.7.2",
    "flax",
    "colorama",
    "tensorflow>=2.20.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "tensorflow-datasets>=4.9.9",
    "orbax-checkpoint", 
]

[project.scripts]
hw04 = "hw04:main"

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"